chore: load dynamically the tile json file
